requirements.txt
pyyaml
numpy
imutils
image
sphinx
pylint
nose
rednose
termcolor
pyserial
